+++ src/stories/HeaderSticky.stories.js
@@ -1,4 +1,6 @@
-// Import mock api data
+// Import mock api data\
+import { computed, onBeforeUnmount, onMounted } from 'vue'
+import { useGlobalStore } from '@/stores/GlobalStore'
 import * as API from '@/stories/mock-api.json'
 import HeaderSticky from '@/lib-components/HeaderSticky'
 
@@ -53,6 +55,47 @@ export function Default() {
         <header-sticky
             :primary-items="primaryItems"
             :secondary-items="parsedSecondaryItems"
+        />
+    `,
+  }
+}
+
+export function FTVAVersion() {
+  return {
+    setup() {
+      onMounted(() => {
+        const globalStore = useGlobalStore()
+
+        const updateWinWidth = () => {
+          console.log(window.innerWidth)
+          globalStore.winWidth = window.innerWidth
+        }
+
+        // Set initial winWidth
+        updateWinWidth()
+
+        window.addEventListener('resize', updateWinWidth)
+
+        // Clean up
+        onBeforeUnmount(() => {
+          window.removeEventListener('resize', updateWinWidth)
+        })
+      })
+    },
+    data() {
+      return {
+        primaryItems,
+      }
+    },
+    provide() {
+      return {
+        theme: computed(() => 'ftva'),
+      }
+    },
+    components: { HeaderSticky },
+    template: `
+        <header-sticky
+            :primary-items="primaryItems"
         />
     `,
   }